+++ spotinst/ocean_aks_strategy/consts.go
@@ -1,0 +1,11 @@
+package ocean_aks_strategy
+
+import (
+	"github.com/spotinst/terraform-provider-spotinst/spotinst/commons"
+)
+
+const (
+	Strategy           commons.FieldName = "strategy"
+	SpotPercentage     commons.FieldName = "spot_percentage"
+	FallbackToOnDemand commons.FieldName = "fallback_to_ondemand"
+)