+++ spotinst/ocean_aks_vm_sizes/consts.go
@@ -1,0 +1,8 @@
+package ocean_aks_vm_sizes
+
+import "github.com/spotinst/terraform-provider-spotinst/spotinst/commons"
+
+const (
+	VMSizes   commons.FieldName = "vm_sizes"
+	Whitelist commons.FieldName = "whitelist"
+)